requirements.txt
# Core
fastapi>=0.128.0
uvicorn>=0.39.0
httpx>=0.28.0
python-dotenv>=1.0.0

# AI
openai>=2.0.0

# WhatsApp / Twilio (fallback)
twilio>=9.0.0

# Google APIs
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.160.0
gspread>=6.1.0

# Scheduling
apscheduler>=3.10.0

# Utils
pandas>=2.0.0
